add completion extendable method
